Separate BuildOperationExecutor and BuildOperationRunner type hierarchies
