+++ platforms/core-runtime/base-services/src/main/java/org/gradle/internal/operations/BuildOperationExecutor.java
@@ -21,6 +21,7 @@ import org.gradle.internal.Factory;
 import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
+import javax.annotation.Nullable;
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
@@ -36,7 +37,7 @@ import javax.annotation.concurrent.ThreadSafe;
  */
 @ThreadSafe
 @ServiceScope(Scopes.BuildSession.class)
-public interface BuildOperationExecutor extends BuildOperationRunner {
+public interface BuildOperationExecutor {
     /**
      * Runs the given build operation synchronously. Invokes the given operation from the current thread.
      *
@@ -44,7 +45,6 @@ public interface BuildOperationExecutor extends BuildOperationRunner {
      * Runtime exceptions are rethrown as is.
      * Checked exceptions are wrapped in {@link BuildOperationInvocationException}.</p>
      */
-    @Override
     void run(RunnableBuildOperation buildOperation);
 
     /**
@@ -55,15 +55,18 @@ public interface BuildOperationExecutor extends BuildOperationRunner {
      * Runtime exceptions are rethrown as is.
      * Checked exceptions are wrapped in {@link BuildOperationInvocationException}.</p>
      */
-    @Override
     <T> T call(CallableBuildOperation<T> buildOperation);
+
+    /**
+     * Executes the given build operation with the given worker, returns the result.
+     */
+    <O extends BuildOperation> void execute(O buildOperation, BuildOperationWorker<O> worker, @Nullable BuildOperationState defaultParent);
 
     /**
      * Starts an operation that can be finished later.
      *
      * When a parent operation is finished any unfinished child operations will be failed.
      */
-    @Override
     BuildOperationContext start(BuildOperationDescriptor.Builder descriptor);
 
     /**
