+++ subprojects/core/src/main/java/org/gradle/internal/operations/DefaultBuildOperationExecutor.java
@@ -24,9 +24,6 @@ import org.gradle.internal.concurrent.ExecutorFactory;
 import org.gradle.internal.concurrent.ManagedExecutor;
 import org.gradle.internal.concurrent.Stoppable;
 import org.gradle.internal.exceptions.DefaultMultiCauseException;
-import org.gradle.internal.logging.progress.ProgressLogger;
-import org.gradle.internal.logging.progress.ProgressLoggerFactory;
-import org.gradle.internal.time.Clock;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -41,23 +38,17 @@ public class DefaultBuildOperationExecutor implements BuildOperationExecutor, St
     private final BuildOperationRunner runner;
     private final BuildOperationQueueFactory buildOperationQueueFactory;
     private final Map<BuildOperationConstraint, ManagedExecutor> managedExecutors = new HashMap<>();
-    private final CurrentBuildOperationRef currentBuildOperationRef = CurrentBuildOperationRef.instance();
+    private final CurrentBuildOperationRef currentBuildOperationRef;
+
     public DefaultBuildOperationExecutor(
-        BuildOperationListener listener,
-        Clock clock,
-        ProgressLoggerFactory progressLoggerFactory,
+        BuildOperationRunner buildOperationRunner,
+        CurrentBuildOperationRef currentBuildOperationRef,
         BuildOperationQueueFactory buildOperationQueueFactory,
         ExecutorFactory executorFactory,
-        ParallelismConfiguration parallelismConfiguration,
-        BuildOperationIdFactory buildOperationIdFactory
+        ParallelismConfiguration parallelismConfiguration
     ) {
-
-        this.runner = new DefaultBuildOperationRunner(
-            currentBuildOperationRef,
-            clock::getCurrentTime,
-            buildOperationIdFactory,
-            () -> new ListenerAdapter(listener, progressLoggerFactory, clock)
-        );
+        this.runner = buildOperationRunner;
+        this.currentBuildOperationRef = currentBuildOperationRef;
         this.buildOperationQueueFactory = buildOperationQueueFactory;
         managedExecutors.put(BuildOperationConstraint.MAX_WORKERS, executorFactory.create("Build operations", parallelismConfiguration.getMaxWorkerCount()));
         managedExecutors.put(BuildOperationConstraint.UNCONSTRAINED, executorFactory.create("Unconstrained build operations", parallelismConfiguration.getMaxWorkerCount() * 10));
@@ -162,60 +153,6 @@ public class DefaultBuildOperationExecutor implements BuildOperationExecutor, St
     public void stop() {
         for (ManagedExecutor pool : managedExecutors.values()) {
             pool.stop();
-        }
-    }
-
-    private static class ListenerAdapter implements DefaultBuildOperationRunner.BuildOperationExecutionListener {
-        private final BuildOperationListener buildOperationListener;
-        private final ProgressLoggerFactory progressLoggerFactory;
-        private final Clock clock;
-        private ProgressLogger progressLogger;
-        private ProgressLogger statusProgressLogger;
-
-        public ListenerAdapter(BuildOperationListener buildOperationListener, ProgressLoggerFactory progressLoggerFactory, Clock clock) {
-            this.buildOperationListener = buildOperationListener;
-            this.progressLoggerFactory = progressLoggerFactory;
-            this.clock = clock;
-        }
-
-        @Override
-        public void start(BuildOperationDescriptor descriptor, BuildOperationState operationState) {
-            buildOperationListener.started(descriptor, new OperationStartEvent(operationState.getStartTime()));
-            ProgressLogger progressLogger = progressLoggerFactory.newOperation(DefaultBuildOperationExecutor.class, descriptor);
-            this.progressLogger = progressLogger.start(descriptor.getDisplayName(), descriptor.getProgressDisplayName());
-        }
-
-        @Override
-        public void progress(BuildOperationDescriptor descriptor, String status) {
-            // Currently, need to start a new progress operation to hold the status, as changing the status of the progress operation replaces the
-            // progress display name on the console, whereas we want to display both the progress display name and the status
-            // This should be pushed down into the progress logger infrastructure so that an operation can have both a display name (that doesn't change) and
-            // a status (that does)
-            if (statusProgressLogger == null) {
-                statusProgressLogger = progressLoggerFactory.newOperation(DefaultBuildOperationExecutor.class, progressLogger);
-                statusProgressLogger.start(descriptor.getDisplayName(), status);
-            } else {
-                statusProgressLogger.progress(status);
-            }
-        }
-
-        @Override
-        public void progress(BuildOperationDescriptor descriptor, long progress, long total, String units, String status) {
-            progress(descriptor, status);
-            buildOperationListener.progress(descriptor.getId(), new OperationProgressEvent(clock.getCurrentTime(), new OperationProgressDetails(progress, total, units)));
-        }
-
-        @Override
-        public void stop(BuildOperationDescriptor descriptor, BuildOperationState operationState, @Nullable BuildOperationState parent, DefaultBuildOperationRunner.ReadableBuildOperationContext context) {
-            if (statusProgressLogger != null) {
-                statusProgressLogger.completed();
-            }
-            progressLogger.completed(context.getStatus(), context.getFailure() != null);
-            buildOperationListener.finished(descriptor, new OperationFinishEvent(operationState.getStartTime(), clock.getCurrentTime(), context.getFailure(), context.getResult()));
-        }
-
-        @Override
-        public void close(BuildOperationDescriptor descriptor, BuildOperationState operationState) {
         }
     }
